backend/requirement.txt
django
twilio
djangorestframework
cloudinary
djangorestframework-jsonapi